+++ src/core/EspressoSystemInterface_cuda.cu
@@ -33,8 +33,8 @@
 // for each property
 
 // Position and charge
-__global__ void split_kernel_rq(CUDA_particle_data *particles, float *r,
-                                float *q, int n) {
+__global__ void split_kernel_rq(CUDA_particle_data *particles, double *r,
+                                double *q, int n) {
   int idx = blockDim.x * blockIdx.x + threadIdx.x;
   if (idx >= n)
     return;
@@ -50,7 +50,7 @@ __global__ void split_kernel_rq(CUDA_particle_data *particles, float *r,
 }
 
 // Charge only
-__global__ void split_kernel_q(CUDA_particle_data *particles, float *q, int n) {
+__global__ void split_kernel_q(CUDA_particle_data *particles, double *q, int n) {
   int idx = blockDim.x * blockIdx.x + threadIdx.x;
   if (idx >= n)
     return;
@@ -63,7 +63,7 @@ __global__ void split_kernel_q(CUDA_particle_data *particles, float *q, int n) {
 }
 
 // Position only
-__global__ void split_kernel_r(CUDA_particle_data *particles, float *r, int n) {
+__global__ void split_kernel_r(CUDA_particle_data *particles, double *r, int n) {
   int idx = blockDim.x * blockIdx.x + threadIdx.x;
   if (idx >= n)
     return;
@@ -79,7 +79,7 @@ __global__ void split_kernel_r(CUDA_particle_data *particles, float *r, int n) {
 
 #ifdef LB_GPU
 // Velocity
-__global__ void split_kernel_v(CUDA_particle_data *particles, float *v, int n) {
+__global__ void split_kernel_v(CUDA_particle_data *particles, double *v, int n) {
   int idx = blockDim.x * blockIdx.x + threadIdx.x;
   if (idx >= n)
     return;
@@ -96,7 +96,7 @@ __global__ void split_kernel_v(CUDA_particle_data *particles, float *v, int n) {
 
 #ifdef DIPOLES
 // Dipole moment
-__global__ void split_kernel_dip(CUDA_particle_data *particles, float *dip,
+__global__ void split_kernel_dip(CUDA_particle_data *particles, double *dip,
                                  int n) {
   int idx = blockDim.x * blockIdx.x + threadIdx.x;
   if (idx >= n)
@@ -113,7 +113,7 @@ __global__ void split_kernel_dip(CUDA_particle_data *particles, float *dip,
 #endif
 
 __global__ void split_kernel_director(CUDA_particle_data *particles,
-                                      float *director, int n) {
+                                      double *director, int n) {
 #ifdef ROTATION
   int idx = blockDim.x * blockIdx.x + threadIdx.x;
   if (idx >= n)
@@ -133,28 +133,28 @@ void EspressoSystemInterface::reallocDeviceMemory(int n) {
   if (m_needsRGpu && ((n != m_gpu_npart) || (m_r_gpu_begin == 0))) {
     if (m_r_gpu_begin != 0)
       cuda_safe_mem(cudaFree(m_r_gpu_begin));
-    cuda_safe_mem(cudaMalloc(&m_r_gpu_begin, 3 * n * sizeof(float)));
+    cuda_safe_mem(cudaMalloc(&m_r_gpu_begin, 3 * n * sizeof(double)));
     m_r_gpu_end = m_r_gpu_begin + 3 * n;
   }
 #ifdef DIPOLES
   if (m_needsDipGpu && ((n != m_gpu_npart) || (m_dip_gpu_begin == 0))) {
     if (m_dip_gpu_begin != 0)
       cuda_safe_mem(cudaFree(m_dip_gpu_begin));
-    cuda_safe_mem(cudaMalloc(&m_dip_gpu_begin, 3 * n * sizeof(float)));
+    cuda_safe_mem(cudaMalloc(&m_dip_gpu_begin, 3 * n * sizeof(double)));
     m_dip_gpu_end = m_dip_gpu_begin + 3 * n;
   }
 #endif
   if (m_needsVGpu && ((n != m_gpu_npart) || (m_v_gpu_begin == 0))) {
     if (m_v_gpu_begin != 0)
       cuda_safe_mem(cudaFree(m_v_gpu_begin));
-    cuda_safe_mem(cudaMalloc(&m_v_gpu_begin, 3 * n * sizeof(float)));
+    cuda_safe_mem(cudaMalloc(&m_v_gpu_begin, 3 * n * sizeof(double)));
     m_v_gpu_end = m_v_gpu_begin + 3 * n;
   }
 
   if (m_needsQGpu && ((n != m_gpu_npart) || (m_q_gpu_begin == 0))) {
     if (m_q_gpu_begin != 0)
       cuda_safe_mem(cudaFree(m_q_gpu_begin));
-    cuda_safe_mem(cudaMalloc(&m_q_gpu_begin, 3 * n * sizeof(float)));
+    cuda_safe_mem(cudaMalloc(&m_q_gpu_begin, 3 * n * sizeof(double)));
     m_q_gpu_end = m_q_gpu_begin + 3 * n;
   }
 
@@ -162,7 +162,7 @@ void EspressoSystemInterface::reallocDeviceMemory(int n) {
       ((n != m_gpu_npart) || (m_director_gpu_begin == 0))) {
     if (m_director_gpu_begin != 0)
       cuda_safe_mem(cudaFree(m_director_gpu_begin));
-    cuda_safe_mem(cudaMalloc(&m_director_gpu_begin, 3 * n * sizeof(float)));
+    cuda_safe_mem(cudaMalloc(&m_director_gpu_begin, 3 * n * sizeof(double)));
     m_director_gpu_end = m_director_gpu_begin + 3 * n;
   }
