+++ src/core/actor/HarmonicWell_cuda.cu
@@ -28,8 +28,8 @@
 #error CU-file includes mpi.h! This should not happen!
 #endif
 
-__global__ void HarmonicWell_kernel(float x, float y, float z, float k, int n,
-                                    float *pos, float *f) {
+__global__ void HarmonicWell_kernel(double x, double y, double z, double k, int n,
+                                    double *pos, double *f) {
 
   int id = blockIdx.x * blockDim.x + threadIdx.x;
 
@@ -41,8 +41,8 @@ __global__ void HarmonicWell_kernel(float x, float y, float z, float k, int n,
   f[3 * id + 2] += k * (z - pos[3 * id + 2]);
 }
 
-void HarmonicWell_kernel_wrapper(float x, float y, float z, float k, int n,
-                                 float *pos, float *f) {
+void HarmonicWell_kernel_wrapper(double x, double y, double z, double k, int n,
+                                 double *pos, double *f) {
   dim3 grid(1, 1, 1);
   dim3 block(1, 1, 1);
